requirements.txt
# AgniRakshak - Main Dependencies
# Install with: pip install -r requirements.txt

# =============================================================================
# WEB FRAMEWORK
# =============================================================================
Flask==2.3.3
Werkzeug==2.3.7

# =============================================================================
# MACHINE LEARNING
# =============================================================================
xgboost==1.7.6
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
joblib==1.3.2

# =============================================================================
# DATA SCIENCE
# =============================================================================
scipy==1.11.1
matplotlib==3.7.2
seaborn==0.12.2

# =============================================================================
# API & HTTP
# =============================================================================
requests==2.31.0
urllib3==2.0.4

# =============================================================================
# ENVIRONMENT & CONFIGURATION
# =============================================================================
python-dotenv==1.0.0

# =============================================================================
# GEOSPATIAL & SATELLITE DATA (OPTIONAL)
# =============================================================================
# Uncomment if you want to use Google Earth Engine
# earthengine-api==0.1.370
# google-auth==2.23.0
# google-auth-oauthlib==1.0.0
# google-cloud-storage==2.10.0

# =============================================================================
# DEVELOPMENT DEPENDENCIES (OPTIONAL)
# =============================================================================
# Uncomment for development
# pytest==7.4.0
# pytest-flask==1.2.0
# black==23.7.0
# flake8==6.0.0
# mypy==1.5.1